Using R and the googlemaps API
The present code shows a Shiny App that interacts with the googlemaps api to get periodic screenshots of the chosen places with the chosen zooms. In the app you can choose the days of the week to run the app, the time frame to run the app, the last day to run the app and there is the possibility to choose a more frequent period for the screenshots.